+++ openradar/config.py
@@ -24,6 +24,7 @@ GROUND_DIR = os.path.join(BUILDOUT_DIR, 'var', 'ground')
 IMG_DIR = os.path.join(BUILDOUT_DIR, 'var', 'img')
 MISC_DIR = os.path.join(BUILDOUT_DIR, 'var', 'misc')
 MULTISCAN_DIR = os.path.join(BUILDOUT_DIR, 'var', 'multiscan')
+NOWCAST_DIR = os.path.join(BUILDOUT_DIR, 'var', 'nowcast')
 RADAR_DIR = os.path.join(BUILDOUT_DIR, 'var', 'radar')
 SOURCE_DIR = os.path.join(BUILDOUT_DIR, 'var', 'source')
 THREDDS_DIR = os.path.join(BUILDOUT_DIR, 'var', 'thredds')
@@ -43,8 +44,8 @@ DECLUTTER_HISTORY = 50
 DECLUTTER_SIZE = 4
 
 # Radar codes
-DWD_RADARS_2011 = ('ase', 'nhb', 'emd')
-DWD_RADARS = ('ess', 'nhb', 'emd')
+START_YEAR = 2013  # expect an id in dwd filenames starting from here
+DWD_RADARS = ('ess', 'nhb', 'emd', 'ase')  # ess is sometimes called ase
 KNMI_RADARS = ('NL60', 'NL61')
 JABBEKE_RADARS = ('JAB',)
 ALL_RADARS = DWD_RADARS + KNMI_RADARS + JABBEKE_RADARS
@@ -62,11 +63,11 @@ RADAR_PATTERNS = [
     re.compile(
         'RAD_(?P<code>.*)_VOL_NA_(?P<timestamp>[0-9]{12})\.h5',
     ),
-    # DWD 2011
+    # DWD archive
     re.compile(
         'raa00-dx_(?P<code>.*)-(?P<timestamp>[0-9]{10})-dwd---bin',
     ),
-    # DWD
+    # DWD operational
     re.compile(
         'raa00-dx_(?P<id>.*)-(?P<timestamp>[0-9]{10})-(?P<code>.*)---bin',
     ),
@@ -90,7 +91,7 @@ TEMPLATE_TIME_JABBEKE = '%Y%m%d%H%M'
 
 # Templates that reveal datetime format when code and id are substituted
 TEMPLATE_KNMI = 'RAD_{code}_VOL_NA_%Y%m%d%H%M.h5'
-TEMPLATE_DWD_2011 = 'raa00-dx_{code}-%y%m%d%H%M-dwd---bin'
+TEMPLATE_DWD_ARCHIVE = 'raa00-dx_{code}-%y%m%d%H%M-dwd---bin'
 TEMPLATE_DWD = 'raa00-dx_{id}-%y%m%d%H%M-{code}---bin'
 TEMPLATE_JABBEKE = '{code}-%Y%m%d%H%M.dBZ.vol.h5'
 TEMPLATE_GROUND = '%Y%m%d%H%M%S_Grondstations_{code}.csv'