+++ openradar/scripts/image.py
@@ -7,24 +7,23 @@ from __future__ import unicode_literals
 from __future__ import absolute_import
 from __future__ import division
 
+from openradar import arguments
 from openradar import config
 from openradar import images
 from openradar import loghelper
-from openradar import utils
-from openradar import scans
 from openradar import products
+from openradar import scans
+from openradar import utils
 
 import argparse
 
 
 def get_image_args():
-    parser = argparse.ArgumentParser(
-        description='Create images from composites.',
+    argument = arguments.Argument()
+    parser = argument.parser(
+        ['image_dir', 'range'],
+        description='Create images from radar data.',
     )
-    parser.add_argument('range',
-                        metavar='RANGE',
-                        type=str,
-                        help='Range for datetimes 20110101-20110103,20110105')
     parser.add_argument('--postfix',
                         type=str,
                         default='',
@@ -38,9 +37,10 @@ def get_image_args():
                         default='r',
                         help='(r)ealtime, (n)ear-realtime or (a)fterwards')
     parser.add_argument('-c', '--product',
-                        choices=['a', 'b', 'c'],
+                        choices=['a', 'b', 'c', 'n'],
                         default='b',
-                        help='(a)ggregate, cali(b)rated or (c)onsistent')
+                        help=('(a)ggregate, cali(b)rated, '
+                              '(c)onsistent or (n)owcast'))
     parser.add_argument('-t', '--timeframe',
                         choices=['f', 'h', 'd'],
                         default='f',
@@ -55,22 +55,27 @@ def get_image_args():
 
 def product_generator(product, prodcode, timeframe, datetimes):
     """ Return product generator. """
-    for datetime in datetimes:
-        if product == 'a':
+    if product == 'a':
+        for datetime in datetimes:
             yield scans.Aggregate(
                 declutter=None,
                 radars=config.ALL_RADARS,
                 datetime=datetime,
                 timeframe=timeframe,
             )
-        if product == 'b':
-            yield products.CalibratedProduct(datetime=datetime,
-                                             prodcode=prodcode,
-                                             timeframe=timeframe)
-        if product == 'c':
-            yield products.ConsistentProduct(datetime=datetime,
-                                             prodcode=prodcode,
-                                             timeframe=timeframe)
+    else:
+        combinations = utils.get_product_combinations(
+            datetimes=datetimes,
+            prodcodes=prodcode,
+            timeframes=timeframe,
+        )
+        Product = dict(
+            b=products.CalibratedProduct,
+            c=products.ConsistentProduct,
+            n=products.NowcastProduct,
+        )[product]
+        for combination in combinations:
+            yield Product(**combination)
 
 
 def main():
@@ -80,16 +85,16 @@ def main():
     # Get products according to args
     args = get_image_args()
 
-    if args['format'] != 'png':
-        raise NotImplementedError('Only png implemented yet.')
-
     multidaterange = utils.MultiDateRange(args['range'])
     products = product_generator(product=args['product'],
                                  prodcode=args['prodcode'],
                                  timeframe=args['timeframe'],
                                  datetimes=multidaterange.iterdatetimes())
 
+    create = dict(png=images.create_png,
+                  tif=images.create_tif)[args['format']]
+
     # Create images with those products
     kwargs = args.copy()
     map(kwargs.pop, ['range', 'product', 'timeframe', 'prodcode'])
-    images.create_png(products, **kwargs)
+    create(products, **kwargs)